requirements.txt
streamlit
langchain
langchain_community
langchain_huggingface
sentence_transformers
PyPDF2
pypdf
pysqlite3-binary
chromadb
